move test helpers to a module
